Refactoring, Same feature, just divide into several files
